+++ src/common/src/tname.c
@@ -39,7 +39,7 @@ char* extractDBName(const char* tableId, char* name) {
   return strncpy(name, &tableId[offset1 + 1], len);
 }
 
-int32_t tableIdPrefix(const char* name, char* prefix, int32_t len) {
+size_t tableIdPrefix(const char* name, char* prefix, int32_t len) {
   tstrncpy(prefix, name, len);
   strcat(prefix, TS_PATH_DELIMITER);
